🏗️ refactor!: restructure folders
