+++ .github/dependabot.yml
@@ -6,51 +6,4 @@ updates:
     interval: daily
     time: "21:00"
   open-pull-requests-limit: 10
-  ignore:
-  - dependency-name: github.com/casbin/casbin/v2
-    versions:
-    - 2.22.0
-    - 2.23.0
-    - 2.23.1
-    - 2.23.2
-    - 2.24.0
-    - 2.24.1
-    - 2.25.2
-    - 2.25.4
-    - 2.26.0
-    - 2.26.1
-    - 2.27.0
-    - 2.28.2
-    - 2.28.3
-  - dependency-name: cloud.google.com/go
-    versions:
-    - 0.75.0
-    - 0.76.0
-    - 0.77.0
-    - 0.81.0
-  - dependency-name: golang.org/x/text
-    versions:
-    - 0.3.5
-    - 0.3.6
-  - dependency-name: github.com/tidwall/sjson
-    versions:
-    - 1.1.4
-    - 1.1.5
-    - 1.1.6
-  - dependency-name: go.mongodb.org/mongo-driver
-    versions:
-    - 1.4.5
-    - 1.4.6
-    - 1.5.1
-  - dependency-name: github.com/google/uuid
-    versions:
-    - 1.2.0
-  - dependency-name: github.com/paulmach/orb
-    versions:
-    - 0.2.1
-  - dependency-name: github.com/stretchr/testify
-    versions:
-    - 1.7.0
-  - dependency-name: github.com/Masterminds/semver/v3
-    versions:
-    - 3.1.1
+
