+++ .github/workflows/go.yml
@@ -27,19 +27,13 @@ jobs:
           # mysql user: "developer" # Required if "mysql root password" is empty, default is empty. The superuser for the specified database. Can use secrets, too
           # mysql password: "" # Required if "mysql user" exists. The password for the "mysql user"
 
-      # - name: Set up Go 1.13
-      #   uses: actions/setup-go@v1
-      #   with:
-      #     go-version: 1.13
-      #   id: go
-
       - name: Check out code into the Go module directory
         uses: actions/checkout@v2
 
-      - name: Set up Go 1.14
+      - name: Set up Go 1.15
         uses: actions/setup-go@v2
         with:
-          go-version: 1.14
+          go-version: 1.15
 
       - name: Get dependencies
         run: |
