+++ Makefile
@@ -1,7 +1,7 @@
-generate-protobuf :
-	@rm -rf ./proto/*.go && \
-	protoc --proto_path=./proto \
-	--go_out=./proto/ --go_opt=paths=source_relative \
-	--go-grpc_out=./proto --go-grpc_opt=paths=source_relative \
-	./proto/*.proto && \
+proto:
+	@rm -rf ./x/proto/*.go && \
+	protoc --proto_path=./x/proto \
+	--go_out=./x/proto/ --go_opt=paths=source_relative \
+	--go-grpc_out=./x/proto --go-grpc_opt=paths=source_relative \
+	./x/proto/*.proto && \
 	echo "proto code generation successful"
