+++ sql/codec/registry.go
@@ -12,7 +12,7 @@ import (
 	"cloud.google.com/go/civil"
 	"github.com/paulmach/orb"
 	"github.com/si3nloong/sqlike/reflext"
-	"github.com/si3nloong/sqlike/spatial"
+	"github.com/si3nloong/sqlike/x/spatial"
 	"golang.org/x/text/currency"
 	"golang.org/x/text/language"
 )
