+++ sql/dialect/mysql/connect.go
@@ -2,7 +2,7 @@ package mysql
 
 import (
 	"github.com/si3nloong/sqlike/sqlike/options"
-	"github.com/si3nloong/sqlike/util"
+	"github.com/si3nloong/sqlike/x/util"
 )
 
 // Connect :
