+++ sql/expr/spatial.go
@@ -3,8 +3,8 @@ package expr
 import (
 	"github.com/paulmach/orb"
 	"github.com/paulmach/orb/encoding/wkt"
-	"github.com/si3nloong/sqlike/spatial"
 	"github.com/si3nloong/sqlike/sqlike/primitive"
+	"github.com/si3nloong/sqlike/x/spatial"
 )
 
 //golint:ignore
